edge/requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.2
kazoo==2.8.0
MarkupSafe==1.1.1
psutil==5.7.3
pytz==2020.1
requests==2.24.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
